Move files for Admin UI pages
