+++ packages/core/package.json
@@ -44,8 +44,8 @@
     "./fields/types/multiselect/views": "./dist/keystone-6-core-fields-types-multiselect-views.js",
     "./fields/types/relationship/views": "./dist/keystone-6-core-fields-types-relationship-views.js",
     "./___internal-do-not-use-will-break-in-patch/artifacts": "./dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-artifacts.js",
-    "./___internal-do-not-use-will-break-in-patch/admin-ui/id-field-view": "./dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-admin-ui-id-field-view.js",
     "./___internal-do-not-use-will-break-in-patch/admin-ui/pages/App": "./dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-admin-ui-pages-App.js",
+    "./___internal-do-not-use-will-break-in-patch/admin-ui/id-field-view": "./dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-admin-ui-id-field-view.js",
     "./___internal-do-not-use-will-break-in-patch/admin-ui/pages/HomePage": "./dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-admin-ui-pages-HomePage.js",
     "./___internal-do-not-use-will-break-in-patch/admin-ui/pages/ItemPage": "./dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-admin-ui-pages-ItemPage.js",
     "./___internal-do-not-use-will-break-in-patch/admin-ui/pages/ListPage": "./dist/keystone-6-core-___internal-do-not-use-will-break-in-patch-admin-ui-pages-ListPage.js",
@@ -146,7 +146,7 @@
       "index.ts",
       "access.ts",
       "___internal-do-not-use-will-break-in-patch/artifacts.ts",
-      "___internal-do-not-use-will-break-in-patch/admin-ui/pages/*/index.tsx",
+      "___internal-do-not-use-will-break-in-patch/admin-ui/pages/*.tsx",
       "___internal-do-not-use-will-break-in-patch/admin-ui/id-field-view.tsx",
       "context.ts",
       "testing/mysql.ts",
