+++ packages/core/src/___internal-do-not-use-will-break-in-patch/admin-ui/pages/CreateItemPage.tsx
@@ -1,0 +1,5 @@
+import { CreateItemPage } from '../../../admin-ui/pages/CreateItemPage/index.tsx'
+
+export const getCreateItemPage = (props: Parameters<typeof CreateItemPage>[0]) => () => (
+  <CreateItemPage {...props} />
+)
