+++ packages/core/src/___internal-do-not-use-will-break-in-patch/admin-ui/pages/HomePage.tsx
@@ -1,0 +1,1 @@
+export { HomePage } from '../../../admin-ui/pages/HomePage/index.tsx'
