+++ packages/core/src/___internal-do-not-use-will-break-in-patch/admin-ui/pages/ItemPage.tsx
@@ -1,0 +1,3 @@
+import { ItemPage } from '../../../admin-ui/pages/ItemPage/index.tsx'
+
+export const getItemPage = (props: Parameters<typeof ItemPage>[0]) => () => <ItemPage {...props} />
