+++ packages/core/src/___internal-do-not-use-will-break-in-patch/admin-ui/pages/ListPage.tsx
@@ -1,0 +1,3 @@
+import { ListPage } from '../../../admin-ui/pages/ListPage/index.tsx'
+
+export const getListPage = (props: Parameters<typeof ListPage>[0]) => () => <ListPage {...props} />
