+++ packages/core/src/___internal-do-not-use-will-break-in-patch/admin-ui/pages/NoAccessPage.tsx
@@ -1,0 +1,5 @@
+import { NoAccessPage } from '../../../admin-ui/pages/NoAccessPage/index.tsx'
+
+export const getNoAccessPage = (props: Parameters<typeof NoAccessPage>[0]) => () => (
+  <NoAccessPage {...props} />
+)
