+++ packages/core/src/admin-ui/pages/ItemPage/common.tsx
@@ -1,4 +1,4 @@
-import { useRouter } from '../../../../admin-ui/router.tsx'
+import { useRouter } from '../../router.tsx'
 import type { HTMLAttributes, ReactNode } from 'react'
 import { Fragment, useMemo, useState } from 'react'
 import isDeepEqual from 'fast-deep-equal'
@@ -12,16 +12,16 @@ import { Grid, HStack } from '@keystar/ui/layout'
 import { breakpointQueries, css, tokenSchema } from '@keystar/ui/style'
 import { toastQueue } from '@keystar/ui/toast'
 import { Heading, Text } from '@keystar/ui/typography'
-import { gql, type TypedDocumentNode, useApolloClient } from '../../../../admin-ui/apollo.ts'
-import { Container, CONTAINER_MAX } from '../../../../admin-ui/components/Container.tsx'
-import { ErrorDetailsDialog } from '../../../../admin-ui/components/Errors.tsx'
-import { ActionDialog } from '../../../../admin-ui/components/ActionDialog.tsx'
+import { gql, type TypedDocumentNode, useApolloClient } from '../../apollo.ts'
+import { Container, CONTAINER_MAX } from '../../components/Container.tsx'
+import { ErrorDetailsDialog } from '../../components/Errors.tsx'
+import { ActionDialog } from '../../components/ActionDialog.tsx'
 import {
   getActionArguments,
   getActionGraphQLArgs,
   getActionGraphQLVariables,
-} from '../../../../admin-ui/utils/actionData.ts'
-import type { ActionMeta, ListMeta } from '../../../../types/index.ts'
+} from '../../utils/actionData.ts'
+import type { ActionMeta, ListMeta } from '../../../types/index.ts'
 
 export function ItemPageHeader({
   list,
