+++ packages/core/src/fields/types/relationship/views/RelationshipTable.tsx
@@ -9,7 +9,7 @@ import { EmptyState } from '../../../../admin-ui/components/EmptyState.tsx'
 import { gql, type TypedDocumentNode, useQuery } from '../../../../admin-ui/apollo.ts'
 import { Text } from '@keystar/ui/typography'
 import { GraphQLErrorNotice } from '../../../../admin-ui/components/index.ts'
-import { PaginationControls } from '../../../../___internal-do-not-use-will-break-in-patch/admin-ui/pages/ListPage/PaginationControls.tsx'
+import { PaginationControls } from '../../../../admin-ui/pages/ListPage/PaginationControls.tsx'
 import type { CountRelationshipValue } from './types.ts'
 import { useRelatedItemHref, useRelatedItemLabel } from './ContextualActions.tsx'
 import { ActionButton } from '@keystar/ui/button'
